# swagger.yaml
---
swagger: "2.0"
info:
  description: This API is the database interface for information related to Twitter.
  version: "1.0.1"
  title: This API is the database interface for information related to Twitter.
  contact:
    email: stanik@informatik.uni-hamburg.de
paths:
  /store/tweet/:
    post:
      summary: Store a list of tweets.
      description: Store a list of tweets.
      operationId: postTweet
      consumes:
        - application/json
      parameters:
        - in: body
          name: tweet
          required: true
          schema:
            $ref: "#/definitions/Tweets"
      responses:
        200:
          description: Tweets successfully stored.
        400:
          description: bad input parameter or no tweet could be retrieved.
  /label/tweet/:
    post:
      summary: stores a human created label for a tweet.
      description: stores a human created label for a tweet.
      operationId: postLabelTwitter
      consumes:
        - application/json
      parameters:
        - in: body
          name: tweetlabel
          required: true
          schema:
            $ref: "#/definitions/TweetLabel"
      responses:
        200:
          description: Tweets successfully stored.
        400:
          description: bad input parameter or no tweet could be retrieved.
  /store/observable/:
    post:
      summary: Stores a twitter profile as under observation (will be crawled frequently)
      description: Stores a twitter profile as under observation (will be crawled frequently)
      operationId: postObservableTwitter
      consumes:
        - application/json
      parameters:
        - in: body
          name: observableTwitter
          required: true
          schema:
            $ref: "#/definitions/ObservableTwitter"
      responses:
        200:
          description: observable twitter profile successfully stored.
        400:
          description: bad input parameter or no tweet could be retrieved.
        500:
          description: could not store the observable in the database.
  /account_name/all:
    get:
      summary: returns a list of strings containing all twitter profile names for which tweets are collected.
      description: returns a list of strings containing all twitter profile names for which tweets are collected.
      operationId: getAllTwitterAccountNames
      produces:
        - application/json
      responses:
        200:
          description: list of twitter profile names.
          schema:
            type: array
            items:
              type: string
  /account_name/{account_name}/class/{tweet_class}:
    get:
      summary: Get all tweets belonging to the given class of the given twitter profile.
      description: Get all tweets belonging to the given class of the given twitter profile.
      operationId: getTweetOfClass
      parameters:
        - name: account_name
          in: path
          description: the unique account name of a twitter user.
          required: true
          type: string
        - name: tweet_class
          in: path
          description: the unique account name of a twitter user.
          required: true
          type: string
      responses:
        200:
          description: list of tweets.
          schema:
            $ref: "#/definitions/Tweets"
        400:
          description: bad input parameter or no tweet could be retrieved.
  /account_name/{account_name}/all:
    get:
      summary: Get all tweets belonging to the given twitter profile.
      description: Get all tweets belonging to the given twitter profile.
      operationId: getAllTweetsOfAccount
      parameters:
        - name: account_name
          in: path
          description: the unique account name of a twitter user.
          required: true
          type: string
      responses:
        200:
          description: list of tweets.
          schema:
            $ref: "#/definitions/Tweets"
        400:
          description: bad input parameter or no tweet could be retrieved.
  /account_name/{account_name}/all/unlabeled:
    get:
      summary: Get all tweets belonging to the given twitter profile that are not labeled by a human.
      description: Get all tweets belonging to the given twitter profile that are not labeled by a human.
      operationId: getAllUnlabeledTweetsOfAccount
      parameters:
        - name: account_name
          in: path
          description: the unique account name of a twitter user.
          required: true
          type: string
      responses:
        200:
          description: list of tweets.
          schema:
            $ref: "#/definitions/Tweets"
        400:
          description: bad input parameter or no tweet could be retrieved.
  /account_name/{account_name}/lang/{lang}/unclassified:
    get:
      summary: Get all tweets belonging to the given twitter profile which had not ben classified yet.
      description: Get all tweets belonging to the given twitter profile which had not ben classified yet.
      operationId: getAllUnclassifiedTweetsOfAccount
      parameters:
        - name: account_name
          in: path
          description: the unique account name of a twitter user.
          required: true
          type: string
        - name: lang
          in: path
          description: the language the tweets are written in.
          required: true
          type: string
      responses:
        200:
          description: list of tweets.
          schema:
            $ref: "#/definitions/Tweets"
        400:
          description: bad input parameter or no tweet could be retrieved.
  /account_name/{account_name}/currentweek:
    get:
      summary: Get all tweets of the current week belonging to the given twitter profile.
      description: Get all tweets of the current week belonging to the given twitter profile.
      operationId: getAllTweetsOfAccountForCurrentWeek
      parameters:
        - name: account_name
          in: path
          description: the unique account name of a twitter user.
          required: true
          type: string
      responses:
        200:
          description: list of tweets.
          schema:
            $ref: "#/definitions/Tweets"
        400:
          description: bad input parameter or no tweet could be retrieved.
  /observables:
    get:
      summary: Get all twitter profiles under observation.
      description: Get all twitter profiles under observation.
      operationId: getObservablesTwitter
      produces:
        - application/json
      responses:
        200:
          description: observable twitter profiles.
          schema:
            $ref: "#/definitions/ObservablesTwitter"
        400:
          description: bad input parameter or no tweet could be retrieved.
    delete:
      summary: Get all twitter profiles under observation.
      description: Get all twitter profiles under observation.
      operationId: deleteObservableTwitter
      consumes:
        - application/json
      parameters:
        - in: body
          name: observable
          required: true
          schema:
            $ref: "#/definitions/ObservableTwitter"
      responses:
        200:
          description: observable successfully deleted.
        400:
          description: bad input parameter or could not delete observable.
definitions:
  Tweets:
    type: array
    items:
      $ref: "#/definitions/Tweet"
  Tweet:
    type: object
    properties:
      created_at:
        type: integer
        example: 20181201
      favorite_count:
        type: integer
        example: 1
      retweet_count:
        type: integer
        example: 1
      status_id:
        type: string
        example: "933476766408200193"
      in_reply_to_screen_name:
        type: string
        example: musteraccount
      lang:
        type: string
        example: en
      text:
        type: string
        example: "@maxmustermannThanks for your message!"
      user_name:
        type: string
        example: maxmustermann
      sentiment:
        type: string
        example: NEUTRAL
      sentiment_score:
        type: integer
        example: -1
      tweet_class:
        type: string
        example: problem_report
  TweetLabel:
    type: object
    properties:
      status_id:
        type: string
        example: "933476766408200193"
      date:
        type: integer
        example: 20181231
      label:
        type: string
        example: "problem_report"
      previous_label:
        type: string
        example: "inquiry"
  ObservablesTwitter:
    type: array
    items:
      $ref: "#/definitions/ObservableTwitter"
  ObservableTwitter:
    type: object
    properties:
      account_name:
        type: string
        example: mustermann
      interval:
        type: string
        example: daily
      lang:
        type: string
        example: en